telegram-ai-worker/requirements.txt
python-telegram-bot==21.11.1
openai>=1.35.0
requests
python-dotenv
